+++ app/src/main/java/com/marco97pa/puntiburraco/TripleFragment.java
@@ -2,18 +2,27 @@ package com.marco97pa.puntiburraco;
 
 import android.Manifest;
 import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.widget.PopupMenu;
 import androidx.fragment.app.Fragment;
+
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.app.ActivityManager;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.ImageDecoder;
 import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
+import android.os.VibrationEffect;
+import android.os.Vibrator;
 import android.preference.PreferenceManager;
 import androidx.annotation.NonNull;
 
@@ -25,28 +34,37 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.snackbar.BaseTransientBottomBar;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.analytics.FirebaseAnalytics;
+import com.yalantis.ucrop.UCrop;
 
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import androidx.core.content.FileProvider;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.animation.LinearInterpolator;
 import android.view.inputmethod.InputMethodManager;
 import android.webkit.WebView;
 import android.widget.CheckBox;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.channels.FileChannel;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
+
+import static android.app.Activity.RESULT_OK;
+import static android.content.Context.ACTIVITY_SERVICE;
 
 /**
  * TRIPLE FRAGMENT
@@ -59,6 +77,8 @@ import java.util.Calendar;
  */
 
 public class TripleFragment extends Fragment {
+
+    public static final String LOG_TAG = "3PlayersFragment";
     int bp1,bp2,bp3, bi1, bi2,bi3, bs1, bs2,bs3,pn1,pn2,pn3,tot1,tot2,tot3,pm1,pm2,pm3, pb1, pb2, pb3;
     int tot=0;
     private TextView textNome1, textNome2, textNome3;
@@ -72,12 +92,24 @@ public class TripleFragment extends Fragment {
     private CheckBox CH1, CH2, CH3;
     private CheckBox RB1, RB2, RB3;
     private CheckBox PZ1, PZ2, PZ3;
+    private ImageView IMG1, IMG2, IMG3; //Images of players
     boolean check1=false, check2=false, check3=false;
     int SET, input_method;
     private AdView mAdView;
     final int PDefault=0;
     boolean win=false;
     String winner,loser1,loser2;
+    //Constants response to import images in app
+    private static int REQUEST_PICTURE_3 = 13;
+    private static int REQUEST_CROP_PICTURE_3 = 23;
+    private static int REQUEST_PICTURE_2 = 12;
+    private static int REQUEST_CROP_PICTURE_2 = 22;
+    private static int REQUEST_PICTURE_1 = 11;
+    private static int REQUEST_CROP_PICTURE_1 = 21;
+    //Constants response to permission request (Android 6.0+)
+    private final static int STORAGE_PERMISSION_PICTURE_1 = 13;
+    private final static int STORAGE_PERMISSION_PICTURE_2 = 23;
+    private final static int STORAGE_PERMISSION_PICTURE_3 = 33;
     private final static int STORAGE_PERMISSION_SCREENSHOT = 30;
 
     public int old_tot1;
@@ -132,6 +164,9 @@ public class TripleFragment extends Fragment {
         PB1 = (EditText) rootView.findViewById(R.id.editPB1);
         PB2 = (EditText) rootView.findViewById(R.id.editPB2);
         PB3 = (EditText) rootView.findViewById(R.id.editPB3);
+        IMG1= (ImageView) rootView.findViewById(R.id.image1);
+        IMG2= (ImageView) rootView.findViewById(R.id.image2);
+        IMG3= (ImageView) rootView.findViewById(R.id.image3); 
 
         sound = MediaPlayer.create(getActivity(), R.raw.fischio);
 
@@ -316,7 +351,9 @@ public class TripleFragment extends Fragment {
                 }
             }
         });
+        
         Restore();
+        
         textNome1.setOnClickListener(new View.OnClickListener() {
 
             @Override
@@ -341,6 +378,265 @@ public class TripleFragment extends Fragment {
                 onSave();
             }
         });
+        
+        IMG1.setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View arg0) {
+                //Animation
+                IMG1.bringToFront();
+                IMG1.invalidate();
+                IMG1.animate()
+                        .scaleX(10)
+                        .scaleY(10)
+                        .setDuration(500)
+                        .setInterpolator(new LinearInterpolator())
+                        .setListener(new AnimatorListenerAdapter() {
+                            @Override
+                            public void onAnimationEnd(Animator animator){
+                                IMG1.setScaleX(1);
+                                IMG1.setScaleY(1);
+                            }
+                        });
+                //Pick image from user Gallery
+                // First, request permission (Android 6.0+ only)
+                if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    requestPermissions(
+                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                            STORAGE_PERMISSION_PICTURE_1);
+                }
+                else {
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_1);
+                }
+
+            }
+        });
+
+        IMG1.setOnLongClickListener(new View.OnLongClickListener() {
+            @Override
+            public boolean onLongClick(View view) {
+                makeVibration();
+                PopupMenu popup = new PopupMenu(getActivity(), view);
+
+                // This activity implements OnMenuItemClickListener
+                popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                    @Override
+                    public boolean onMenuItemClick(MenuItem item) {
+                        switch (item.getItemId()) {
+                            case R.id.removeText:
+                                textNome1.setText(getString(R.string.g1));
+                                onSave();
+                                return true;
+                            case R.id.removeImage:
+                                File file1 = new File(getActivity().getFilesDir(), "img_m3_1.jpg");
+                                file1.delete();
+                                IMG1.setImageResource(R.drawable.circle_placeholder);
+                                return true;
+                            default:
+                                return false;
+                        }
+                    }
+                });
+                popup.inflate(R.menu.actions);
+                popup.show();
+                return false;
+            }
+
+        });
+
+        IMG2.setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View arg0) {
+                //Animation
+                IMG2.bringToFront();
+                IMG2.invalidate();
+                IMG2.animate()
+                        .scaleX(10)
+                        .scaleY(10)
+                        .setDuration(500)
+                        .setInterpolator(new LinearInterpolator())
+                        .setListener(new AnimatorListenerAdapter() {
+                            @Override
+                            public void onAnimationEnd(Animator animator){
+                                IMG2.setScaleX(1);
+                                IMG2.setScaleY(1);
+                            }
+                        });
+                //Pick image from user Gallery
+                // First, request permission
+                if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    requestPermissions(
+                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                            STORAGE_PERMISSION_PICTURE_2);
+                }
+                else {
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_2);
+                }
+
+            }
+        });
+
+        IMG2.setOnLongClickListener(new View.OnLongClickListener() {
+            @Override
+            public boolean onLongClick(View view) {
+                makeVibration();
+                PopupMenu popup = new PopupMenu(getActivity(), view);
+
+                // This activity implements OnMenuItemClickListener
+                popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                    @Override
+                    public boolean onMenuItemClick(MenuItem item) {
+                        switch (item.getItemId()) {
+                            case R.id.removeText:
+                                textNome2.setText(getString(R.string.g2));
+                                onSave();
+                                return true;
+                            case R.id.removeImage:
+                                File file2 = new File(getActivity().getFilesDir(), "img_m3_2.jpg");
+                                file2.delete();
+                                IMG2.setImageResource(R.drawable.circle_placeholder);
+                                return true;
+                            default:
+                                return false;
+                        }
+                    }
+                });
+                popup.inflate(R.menu.actions);
+                popup.show();
+                return false;
+            }
+
+        });
+
+        IMG3.setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View arg0) {
+                //Animation
+                IMG3.bringToFront();
+                IMG3.invalidate();
+                IMG3.animate()
+                        .scaleX(10)
+                        .scaleY(10)
+                        .setDuration(500)
+                        .setInterpolator(new LinearInterpolator())
+                        .setListener(new AnimatorListenerAdapter() {
+                            @Override
+                            public void onAnimationEnd(Animator animator){
+                                IMG3.setScaleX(1);
+                                IMG3.setScaleY(1);
+                            }
+                        });
+                //Pick image from user Gallery
+                // First, request permission
+                if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    requestPermissions(
+                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                            STORAGE_PERMISSION_PICTURE_3);
+                }
+                else {
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_3);
+                }
+
+            }
+        });
+
+        IMG3.setOnLongClickListener(new View.OnLongClickListener() {
+            @Override
+            public boolean onLongClick(View view) {
+                makeVibration();
+                PopupMenu popup = new PopupMenu(getActivity(), view);
+
+                // This activity implements OnMenuItemClickListener
+                popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+                    @Override
+                    public boolean onMenuItemClick(MenuItem item) {
+                        switch (item.getItemId()) {
+                            case R.id.removeText:
+                                textNome3.setText(getString(R.string.g3));
+                                onSave();
+                                return true;
+                            case R.id.removeImage:
+                                File file3 = new File(getActivity().getFilesDir(), "img_m3_3.jpg");
+                                file3.delete();
+                                IMG3.setImageResource(R.drawable.circle_placeholder);
+                                return true;
+                            default:
+                                return false;
+                        }
+                    }
+                });
+                popup.inflate(R.menu.actions);
+                popup.show();
+                return false;
+            }
+
+        });
+
+        /**
+         * RECOVER IMAGES OF LAST GAME
+         * Images are showed only if user has choose it in settings
+         */
+        //On lowRamDevice images are disabled par default
+        boolean isLowRamDevice = false;
+        ActivityManager am = (ActivityManager) getActivity().getSystemService(ACTIVITY_SERVICE);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+            isLowRamDevice = am.isLowRamDevice();
+            Log.d(LOG_TAG, "isLowRamDevice? " + Boolean.toString(isLowRamDevice));
+        }
+        
+        Boolean isImgActivated = sharedPref.getBoolean("img", !isLowRamDevice) ;
+        if(isImgActivated) {
+            //BitmapFactory -> ImageDecoder per Android 9.0+ P fix
+            Bitmap bitmap1 = null, bitmap2 = null, bitmap3 = null;
+            if(android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.P){
+                try {
+                    File bmp1 = new File(getActivity().getFilesDir() + "/img_m3_1.jpg");
+                    ImageDecoder.Source source1 = ImageDecoder.createSource(bmp1);
+                    bitmap1 = ImageDecoder.decodeBitmap(source1);
+                } catch (IOException e) { e.printStackTrace(); }
+                try {
+                    File bmp2 = new File(getActivity().getFilesDir() + "/img_m3_2.jpg");
+                    ImageDecoder.Source source2 = ImageDecoder.createSource(bmp2);
+                    bitmap2 = ImageDecoder.decodeBitmap(source2);
+                } catch (IOException e) { e.printStackTrace(); }
+                try {
+                    File bmp3 = new File(getActivity().getFilesDir() + "/img_m3_3.jpg");
+                    ImageDecoder.Source source3 = ImageDecoder.createSource(bmp3);
+                    bitmap3 = ImageDecoder.decodeBitmap(source3);
+                } catch (IOException e) { e.printStackTrace(); }
+            }
+            else{
+                bitmap1 = BitmapFactory.decodeFile(getActivity().getFilesDir() + "/img_m3_1.jpg");
+                bitmap2 = BitmapFactory.decodeFile(getActivity().getFilesDir()+"/img_m3_2.jpg");
+                bitmap3 = BitmapFactory.decodeFile(getActivity().getFilesDir()+"/img_m3_3.jpg");
+            }
+            //Set Bitmap to Image if not null
+            if(bitmap1 != null) {
+                IMG1.setImageBitmap(bitmap1);
+            }
+            if(bitmap2 != null) {
+                IMG2.setImageBitmap(bitmap2);
+            }
+            if(bitmap3 != null) {
+                IMG3.setImageBitmap(bitmap3);
+            }
+        }
+        else{
+            IMG1.setVisibility(View.GONE);
+            IMG2.setVisibility(View.GONE);
+            IMG3.setVisibility(View.GONE);
+        }
+        //improves usability in Android N with MultiWindow and avoids bugs
+        if(android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            if (getActivity().isInMultiWindowMode()) {
+                IMG1.setVisibility(View.GONE);
+                IMG2.setVisibility(View.GONE);
+                IMG3.setVisibility(View.GONE);
+                Log.d(LOG_TAG, "images disabled: isInMultiWindowMode = true");
+            }
+        }
 
         return rootView;
     }
@@ -493,6 +789,16 @@ public class TripleFragment extends Fragment {
         textNome2.setText(getString(R.string.g2));
         textNome3.setText(getString(R.string.g3));
         onSave();
+        //reset Immagini
+        File file1 = new File(getActivity().getFilesDir(), "img_m3_1.jpg");
+        file1.delete();
+        File file2 = new File(getActivity().getFilesDir(), "img_m3_2.jpg");
+        file2.delete();
+        File file3 = new File(getActivity().getFilesDir(), "img_m3_3.jpg");
+        file2.delete();
+        IMG1.setImageResource(R.drawable.circle_placeholder);
+        IMG2.setImageResource(R.drawable.circle_placeholder);
+        IMG3.setImageResource(R.drawable.circle_placeholder);
     }
 
     //AVVIA RESET
@@ -969,6 +1275,56 @@ public class TripleFragment extends Fragment {
     @Override
     public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
         switch (requestCode) {
+            case STORAGE_PERMISSION_PICTURE_1:
+            {
+                // If request is cancelled, the result arrays are empty.
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+
+                    // permission was granted, yay!
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_1);
+
+                } else {
+
+                    // permission denied, boo!
+                    Toast.makeText(getActivity(), getString(R.string.marshmallow_alert), Toast.LENGTH_LONG).show();
+
+                }
+                return;
+            }
+
+            case STORAGE_PERMISSION_PICTURE_2:
+            {
+                // If request is cancelled, the result arrays are empty.
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+
+                    // permission was granted, yay!
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_2);
+
+                } else {
+
+                    // permission denied, boo!
+                    Toast.makeText(getActivity(), getString(R.string.marshmallow_alert), Toast.LENGTH_LONG).show();
+
+                }
+                return;
+            }
+
+            case STORAGE_PERMISSION_PICTURE_3:
+            {
+                // If request is cancelled, the result arrays are empty.
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+
+                    // permission was granted, yay!
+                    startActivityForResult(MediaStoreUtils.getPickImageIntent(getActivity()), REQUEST_PICTURE_3);
+
+                } else {
+
+                    // permission denied, boo!
+                    Toast.makeText(getActivity(), getString(R.string.marshmallow_alert), Toast.LENGTH_LONG).show();
+
+                }
+                return;
+            }
 
             case STORAGE_PERMISSION_SCREENSHOT:
             {
@@ -1154,5 +1510,124 @@ public class TripleFragment extends Fragment {
         return data;
     }
 
+    /**
+     * ON ACTIVITY RESULT
+     * Sets images choosen by user and manages its crop
+     */
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if ((requestCode == REQUEST_PICTURE_1) && (resultCode == RESULT_OK)) {
+            // When the user is done picking a picture, let's start the CropImage Activity,
+            Uri photo = data.getData();
+
+            File file = new File(getActivity().getFilesDir(), "img_m3_1.jpg");
+            Uri result = Uri.fromFile(file);
+            startActivityForResult(
+                    UCrop.of(photo, result)
+                            .withAspectRatio(1, 1)
+                            .withMaxResultSize(1080, 1080)
+                            .getIntent(getActivity()
+                            ), REQUEST_CROP_PICTURE_1);
+        }
+
+        if (resultCode == RESULT_OK && requestCode == REQUEST_CROP_PICTURE_1) {
+            // When we are done cropping, display it in the ImageView.
+            final Uri resultUri = UCrop.getOutput(data);
+            ImageView IMG1 = (ImageView) getView().findViewById(R.id.image1);
+            IMG1.setImageURI(resultUri);
+
+        } else if (resultCode == UCrop.RESULT_ERROR) {
+            final Throwable cropError = UCrop.getError(data);
+        }
+
+
+        if ((requestCode == REQUEST_PICTURE_2) && (resultCode == RESULT_OK)) {
+            // When the user is done picking a picture, let's start the CropImage Activity,
+            Uri photo = data.getData();
+
+            File file = new File(getActivity().getFilesDir(), "img_m3_2.jpg");
+            Uri result = Uri.fromFile(file);
+            startActivityForResult(
+                    UCrop.of(photo, result)
+                            .withAspectRatio(1, 1)
+                            .withMaxResultSize(1080, 1080)
+                            .getIntent(getActivity()
+                            ), REQUEST_CROP_PICTURE_2);
+        }
+
+        if (resultCode == RESULT_OK && requestCode == REQUEST_CROP_PICTURE_2) {
+            // When we are done cropping, display it in the ImageView.
+            final Uri resultUri = UCrop.getOutput(data);
+            ImageView IMG2 = (ImageView) getView().findViewById(R.id.image2);
+            IMG2.setImageURI(resultUri);
+
+        } else if (resultCode == UCrop.RESULT_ERROR) {
+            final Throwable cropError = UCrop.getError(data);
+        }
+
+        if ((requestCode == REQUEST_PICTURE_3) && (resultCode == RESULT_OK)) {
+            // When the user is done picking a picture, let's start the CropImage Activity,
+            Uri photo = data.getData();
+
+            File file = new File(getActivity().getFilesDir(), "img_m3_3.jpg");
+            Uri result = Uri.fromFile(file);
+            startActivityForResult(
+                    UCrop.of(photo, result)
+                            .withAspectRatio(1, 1)
+                            .withMaxResultSize(1080, 1080)
+                            .getIntent(getActivity()
+                            ), REQUEST_CROP_PICTURE_3);
+        }
+
+        if (resultCode == RESULT_OK && requestCode == REQUEST_CROP_PICTURE_3) {
+            // When we are done cropping, display it in the ImageView.
+            final Uri resultUri = UCrop.getOutput(data);
+            ImageView IMG3 = (ImageView) getView().findViewById(R.id.image3);
+            IMG3.setImageURI(resultUri);
+
+        } else if (resultCode == UCrop.RESULT_ERROR) {
+            final Throwable cropError = UCrop.getError(data);
+        }
+
+    }
+
+    /**
+     * SAVE IMAGE
+     * Saves profile images in application data path.
+     * It is invoked when setting a new profile image
+     */
+    public void saveImage(File file, File croppedImageFile){
+        try {
+            FileInputStream inStream = new FileInputStream(croppedImageFile);
+            FileOutputStream outStream = new FileOutputStream(file);
+            FileChannel inChannel = inStream.getChannel();
+            FileChannel outChannel = outStream.getChannel();
+            inChannel.transferTo(0, inChannel.size(), outChannel);
+            inStream.close();
+            outStream.close();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * MAKE VIBRATION
+     */
+    public void makeVibration(){
+        Vibrator vibrator = (Vibrator) getActivity().getSystemService(Context.VIBRATOR_SERVICE);
+        if (vibrator.hasVibrator()) {
+            if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                VibrationEffect effect = VibrationEffect.createOneShot(50, VibrationEffect.DEFAULT_AMPLITUDE);
+                vibrator.vibrate(effect);
+            }
+            else{
+                vibrator.vibrate(50);
+            }
+        }
+    }
 
 }